homework-lesson-5/pyproject.toml
[project]
name = "homework-lesson-5"
version = "0.1.0"
description = "RAG research agent (lesson 5)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "ddgs>=7.0",
    "langchain>=1.2.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "llama-index-core>=0.12.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-retrievers-bm25>=0.4.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "trafilatura>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
